refactor Myaccount comp
