+++ package.json
@@ -30,6 +30,7 @@
     "react-dom": "^19.1.1",
     "react-redux": "^9.2.0",
     "react-router-dom": "^7.9.1",
+    "reactflow": "^11.11.4",
     "redux-persist": "^6.0.0"
   },
   "devDependencies": {
