+++ src/pages/Auth/LoginPage.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import { useNavigate, Link, useLocation } from 'react-router-dom'
-import { useAppDispatch, useAppSelector } from '../hooks/redux'
+import { useAppDispatch, useAppSelector } from '../../hooks/redux'
 import { loginUser, googleLogin, clearError } from '@/stores/slices/authSlice'
 import GoogleSignInButton from '@/components/ui/GoogleSignInButton'
 import { Users, Eye, EyeOff } from 'lucide-react'
