+++ src/pages/Auth/RegisterPage.tsx
