+++ src/pages/MyAccount/FamilyTree.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+const FamilyTree: React.FC = () => {
+  return (
+    <div className="bg-white shadow-sm rounded-lg border border-gray-200">
+      <div className="p-8">
+        <div className="text-center py-12">
+          <h3 className="text-xl font-semibold text-gray-600 mb-2">
+            Thông tin gia phả
+          </h3>
+          <p className="text-gray-500">Chức năng đang được phát triển</p>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default FamilyTree;
