+++ src/routes/AppRouter.tsx
@@ -1,16 +1,16 @@
 import { Navigate, createBrowserRouter, RouterProvider } from 'react-router-dom'
 import { ProtectedRoute } from './ProtectedRoute'
 // import { PERMISSIONS } from '../utils/permissions'
-import LoginPage from '@/pages/LoginPage'
-import RegisterPage from '@/pages/RegisterPage'
+import LoginPage from '@/pages/Auth/LoginPage'
+import RegisterPage from '@/pages/Auth/RegisterPage'
 import UnauthorizedPage from '@/components/shared/UnauthorizedPage'
 import NotFoundPage from '@/components/shared/NotFoundPage'
 import HomePage from '@/pages/HomePage'
-import DashboardPage from '@/pages/DashboardPage'
-import FamilyTreePage from '@/pages/FamilyTreePage'
-import AdminPage from '@/pages/AdminPage'
+import FamilyTreePage from '@/pages/FamilytreeList/FamilyTreePage'
+import AdminPage from '@/pages/Admin/AdminPage'
 import AuthLayout from '@/components/layout/AuthLayout'
 import MainLayout from '@/components/layout/MainLayout'
+import MyAccountPage from '@/pages/MyAccount/MyAccountPage'
 
 const router = createBrowserRouter([
   // Public routes
@@ -66,7 +66,7 @@ const router = createBrowserRouter([
     path: '/dashboard',
     element: (
       <MainLayout>
-        <DashboardPage />
+        <MyAccountPage />
       </MainLayout>
     ),
   },
